Quan Wen's centroid tracker
This calculates the centroid of the entire binary image and uses that for tracking. There is no boundary or contour finding or centerline calculations.

A few lines of code are used here with permission from Quan Wen's single_neuron_tracking repository which in turn was based on an older version of the MindControl software.

This code has not been tested for function nor for compilation.